+++ app/src/main/AndroidManifest.xml
@@ -22,10 +22,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="frgp.utn.edu.ar.quepasa.presentation.activity.auth.SampleActivity"
-            android:label="Testing Activity"
-            android:exported="true"/>
-        <activity
             android:exported="true"
             android:name="frgp.utn.edu.ar.quepasa.presentation.activity.auth.AuthenticatedActivity" />
         <activity
